Add content management hooks and refactor content store
- Introduced new hooks for managing content data, platform-specific content, and content operations, enhancing modularity and reusability.
- Added content processing functions for various platforms (YouTube, Instagram, Gmail) to streamline data transformation.
- Refactored the content store to utilize new types and improve state management, including loading and error handling.
- Updated imports to reflect the new structure, ensuring consistency across the codebase.
